+++ changelog.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 9.14.0
+
+* `[strutil]` Added new method `ReadField` for reading space/tab separated fields from given data
+* `[system]` Code refactoring
+
 ### 9.13.0
 
 * `[system]` Improved CPU usage calculation